pyproject.toml
[project]
name = "telcenter-core-ai-agent"
requires-python = ">=3.12"
version = "0.1.0"
dependencies = [
    "pika>=1.3.2",
    "python-dotenv>=1.2.1",
    "google-generativeai>=0.8.3",
    "requests>=2.32.3",
]
